+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 ### Enhancements
 * New subcommand:
   - `hybran compare`: compare two annotations of the same genome
-* New option `-r`/`--ref-names-only` for `hybran standardize`.
+* Enhancements to `hybran standardize`:
+  - New option `-r`/`--ref-names-only`.
+  - Simplified usage: If passing in a hybran output directory, no other arguments will be necessary.
 * Annotations now include pseudoscan evidence codes as note fields.
 These were previously only found in the pseudoscan report files.